feat: enhance media image generation with new utility functions and Deno configuration
