+++ .vscode/settings.json
@@ -6,11 +6,21 @@
     "editor.defaultFormatter": "biomejs.biome"
   },
   "[typescriptreact]": {
-    "editor.defaultFormatter": "biomejs.biome"
+    "editor.defaultFormatter": "esbenp.prettier-vscode"
   },
   "javascript.preferences.importModuleSpecifier": "non-relative",
   "typescript.preferences.importModuleSpecifier": "non-relative",
   "files.associations": {
-    "*.css": "tailwindcss"
-  }
+    "*.css": "tailwindcss",
+    "supabase/functions/**/*.ts": "typescript",
+    "supabase/functions/**/*.tsx": "typescriptreact"
+  },
+  // Deno settings for Edge Functions
+  "deno.enable": true,
+  "deno.lint": true,
+  "deno.enablePaths": [
+    "supabase/functions"
+  ],
+  // Use Deno language server for Edge Functions
+  "deno.config": "./supabase/**/deno.json",
 }
