+++ biome.json
@@ -13,7 +13,8 @@
       "**/.next/**",
       "**/build/**",
       "coverage/**",
-      "public/**"
+      "public/**",
+      "supabase/functions/**"
     ]
   },
   "formatter": {
